migrate babel_parser to ts and correct folder and update lint error
