+++ package.json
@@ -23,12 +23,12 @@
     "prepublish": "yarn run build",
     "test": "jest",
     "flow": "flow",
-    "lint": "eslint \"?(__mocks__|src|tests)/**/*.+(js|ts)\" fixtures/parser_tests.js",
+    "lint": "eslint \"?(__mocks__|src|tests)/**/*.+(js|ts)\" ",
     "lint:fix": "eslint \"?(__mocks__|src|tests)/**/*.+(js|ts)\" --fix",
     "ci": "yarn type-check && yarn lint && yarn test --coverage && yarn flow",
-    "prettier": "prettier --check \"?(__mocks__|src|tests)/**/*.+(js|ts)\" fixtures/parser_tests.js",
+    "prettier": "prettier --check \"?(__mocks__|src|tests)/**/*.+(js|ts)\" ",
     "prettier-write": "prettier --write",
-    "prettier-project": "yarn prettier-write \"?(__mocks__|src|tests)/**/*.+(js|ts)\" fixtures/parser_tests.js",
+    "prettier-project": "yarn prettier-write \"?(__mocks__|src|tests)/**/*.+(js|ts)\" ",
     "type-check": "tsc --noEmit",
     "type-check:watch": "yarn type-check -- --watch"
   },
@@ -37,7 +37,8 @@
     "@babel/core": "^7.6.2",
     "@babel/preset-env": "^7.6.2",
     "@babel/preset-flow": "^7.0.0",
-    "@babel/preset-typescript": "^7.8.3",
+    "@babel/preset-typescript": "^7.13.0",
+    "@types/jest": "^26.0.23",
     "@typescript-eslint/eslint-plugin": "^2.18.0",
     "@typescript-eslint/parser": "^2.18.0",
     "babel-core": "7.0.0-bridge.0",
@@ -57,7 +58,7 @@
     "flow-bin": "^0.97.0",
     "jest": "^26.6.3",
     "lint-staged": "^8.1.5",
-    "prettier": "^1.17.0",
+    "prettier": "^2.3.0",
     "typescript": "^3.8.2"
   },
   "dependencies": {
